requirements.txt
django==3.2
pytest==7.2.0
python-dotenv==0.21
psycopg2==2.9.4
django-split-settings==1.2.0
django-debug-toolbar==3.4.0
pendulum==2.1.2
sqlalchemy==1.4.42
pandas==1.5.1